Add .template extension to template files
The template files do not contain valid syntax...yet. Therefore it is not appropriate to give them .h/.py extensions.
The primary motivation was found when byte-compiling python code automatically. To do this in Fedora, the script searches for files ending in .py and tries to bytecompile them, which fails on the python template file with its current name.