Revert #140
We discovered that there is an issue when building overlays with this in
place.  This didn't show up in the normal CI process where we don't
overlay.  The `SYSTEM` parameter causes gcc to use `-isystem`, which
does suppress warnings as the PR intended, but it also re-orders the
include paths, which can cause overlays to break.

Signed-off-by: Michael Carroll <michael@openrobotics.org>